+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.24.0
+
+### Features
+
+- Added method `getNextMessages` to get the next batch of messages for the dialog
+- Implemented new statuses "not-authorized" and "error" for `chatStatus` state
+
 ## 0.23.2
 
 ### Misc
@@ -23,7 +30,7 @@
 
 ### Features
 
-- Implemented `chatStatus` state with "disconnected", "connecting" and "connected"
+- Implemented `chatStatus` state with "disconnected", "connecting" and "connected" statuses
 - Implemented `processOnMessageSent` function to handle actions triggered by lost/sent message
 
 ## 0.22.0